BAGEL: Hopefully fix CRect TopLeft/BottomRight casts. Again.
This reverts revert commit f85a43163ee64e7244e5eb5f4de72396d03036b8.

+++ engines/bagel/mfc/wingdi.h
@@ -654,12 +654,12 @@ typedef struct tagBITMAPINFO {
 
 /* Bitmap Header Definition */
 typedef struct tagBITMAP {
-	long        bmType;
-	long        bmWidth;
-	long        bmHeight;
-	long        bmWidthBytes;
-	uint16        bmPlanes;
-	uint16        bmBitsPixel;
+	int bmType;
+	int bmWidth;
+	int bmHeight;
+	int bmWidthBytes;
+	uint16 bmPlanes;
+	uint16 bmBitsPixel;
 	void *     bmBits;
 } BITMAP, *PBITMAP, NEAR *NPBITMAP, FAR *LPBITMAP;
